hh: rename factory module
